git-debimport bluez_5.43.orig.tar.gz

+++ obexd/plugins/bluetooth.c
@@ -34,18 +34,20 @@
 #include <sys/socket.h>
 
 #include <glib.h>
-#include <gdbus/gdbus.h>
 
-#include "btio/btio.h"
+#include "lib/bluetooth.h"
 #include "lib/uuid.h"
 
-#include "obexd.h"
-#include "plugin.h"
-#include "server.h"
-#include "obex.h"
-#include "transport.h"
-#include "service.h"
-#include "log.h"
+#include "gdbus/gdbus.h"
+
+#include "btio/btio.h"
+#include "obexd/src/obexd.h"
+#include "obexd/src/plugin.h"
+#include "obexd/src/server.h"
+#include "obexd/src/obex.h"
+#include "obexd/src/transport.h"
+#include "obexd/src/service.h"
+#include "obexd/src/log.h"
 
 #define BT_RX_MTU 32767
 #define BT_TX_MTU 32767
@@ -73,8 +75,8 @@ static void connect_event(GIOChannel *io, GError *err, void *user_data)
 	struct bluetooth_profile *profile = user_data;
 	struct obex_server *server = profile->server;
 	int type;
-	int omtu = BT_TX_MTU;
-	int imtu = BT_RX_MTU;
+	uint16_t omtu = BT_TX_MTU;
+	uint16_t imtu = BT_RX_MTU;
 	gboolean stream = TRUE;
 	socklen_t len = sizeof(int);
